user_management/requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.0
typing-extensions==4.12.2
whitenoise==6.7.0
psycopg
inflection
uritemplate
daphne
redis
channels
channels-redis